+++ lib/src/common/constants.dart
@@ -1,2 +1,29 @@
+// API
+// ignore_for_file: constant_identifier_names
+
+const String API_URL = "https://reqres.in/api";
+
+const String API_USERS_TOTAL = "total";
+const String API_USERS_PER_PAGE = "per_page";
+const String API_USERS_DATA = "data";
+
+const String API_RESOURCES_TOTAL = "total";
+const String API_RESOURCES_PER_PAGE = "per_page";
+const String API_RESOURCES_DATA = "data";
+
+// User
+const String API_USER_ID = "id";
+const String API_USER_EMAIL = "email";
+const String API_USER_FIRSTNAME = "first_name";
+const String API_USER_LASTNAME = "last_name";
+const String API_USER_AVATAR = "avatar";
+
+// Resource
+const String API_RESOURCE_ID = "id";
+const String API_RESOURCE_NAME = "name";
+const String API_RESOURCE_COLOR = "color";
+const String API_RESOURCE_YEAR = "year";
+const String API_RESOURCE_PANTONE_VALUE = "pantone_values";
+
 // App Title
 const String APP_TITLE = "SWD Nucleus Tech Team - ReqRes consumer";